+++ retrowave_downloader/src/background/background.js
@@ -1,18 +1,45 @@
-chrome.runtime.onInstalled.addListener(function () {
-    chrome.declarativeContent.onPageChanged.removeRules(undefined, function () {
-        chrome.declarativeContent.onPageChanged.addRules([
-            {
-                conditions: [
-                    new chrome.declarativeContent.PageStateMatcher({
-                        pageUrl: {hostEquals: 'retrowave.ru', schemes: ['http', 'https']},
-                    })
-                ],
-                actions: [new chrome.declarativeContent.ShowPageAction()]
-            }
-        ]);
-    });
-});
+class RetrowaveDownloader {
 
-chrome.pageAction.onClicked.addListener(function (tab) {
-    chrome.tabs.sendMessage(tab.id, {type:'downloadCurrentPlaying'});
-});
+    constructor() {
+        const self = this;
+
+        self.url_query = '*://*.retrowave.ru/*';
+
+        self.update_connection();
+        self.setup_listeners()
+    }
+
+    setup_listeners() {
+        const self = this;
+
+        chrome.browserAction.onClicked.addListener(function (tab) {
+            self.send_download_action.call(self);
+        });
+        chrome.tabs.onCreated.addListener(function (tab) {
+            self.update_connection.call(self);
+        });
+        chrome.tabs.onUpdated.addListener(function (tab) {
+            self.update_connection.call(self);
+        });
+        chrome.tabs.onRemoved.addListener(function (tab) {
+            self.update_connection.call(self);
+        });
+    }
+
+    update_connection() {
+        const self = this;
+
+        chrome.tabs.query({'url': self.url_query}, function (tabs) {
+            tabs.length > 0 ? self.tab = tabs[0] : self.tab = undefined;
+        });
+    }
+
+    send_download_action() {
+        const self = this;
+
+        if (!self.tab) return;
+        chrome.tabs.sendMessage(self.tab.id, {type: 'downloadCurrentPlaying'});
+    }
+}
+
+let retrowave_downloader = new RetrowaveDownloader();